Add a post-generate extension point for bootstraps.
