+++ src/briefcase/bootstraps/__init__.py
@@ -1,5 +1,6 @@
 from briefcase.bootstraps.base import BaseGuiBootstrap  # noqa: F401
 from briefcase.bootstraps.console import ConsoleBootstrap  # noqa: F401
+from briefcase.bootstraps.empty import EmptyBootstrap  # noqa: F401
 from briefcase.bootstraps.pygame import PygameGuiBootstrap  # noqa: F401
 from briefcase.bootstraps.pyside6 import PySide6GuiBootstrap  # noqa: F401
 from briefcase.bootstraps.toga import TogaGuiBootstrap  # noqa: F401
