+++ tests/commands/convert/test_input_app_type.py
@@ -1,18 +1,6 @@
 import pytest
 
 
-def test_overrides_are_used_for_console(convert_command):
-    overrides = {"console_app": "Console"}
-    out = convert_command.build_gui_context({}, overrides)
-    assert out == {"console_app": True, "gui_framework": "None"}
-
-
-def test_overrides_are_used_for_GUI(convert_command):
-    overrides = {"console_app": "GUI"}
-    out = convert_command.build_gui_context({}, overrides)
-    assert out == {"console_app": False, "gui_framework": "None"}
-
-
 @pytest.mark.parametrize(
     "input, result",
     [
@@ -29,5 +17,5 @@ def test_overrides_are_used_for_GUI(convert_command):
 def test_app_type(convert_command, input, result):
     """The user can be asked for the app type."""
     convert_command.input.values = input
-    out = convert_command.input_console_app(None)
+    out = convert_command.input_app_type(None)
     assert out == result
