+++ tests/commands/new/test_new_app.py
@@ -6,6 +6,7 @@ from cookiecutter import exceptions as cookiecutter_exceptions
 from cookiecutter.main import cookiecutter
 
 import briefcase
+from briefcase.bootstraps import BaseGuiBootstrap
 from briefcase.commands import NewCommand
 from briefcase.console import Console, Log
 from briefcase.exceptions import (
@@ -52,6 +53,8 @@ def test_new_app(
         "app_name": "myapplication",
     }
     new_command.build_app_context = mock.MagicMock(return_value=app_context)
+    bootstrap = BaseGuiBootstrap(new_command.logger, new_command.input, {})
+    new_command.create_bootstrap = mock.MagicMock(return_value=bootstrap)
     new_command.build_gui_context = mock.MagicMock(
         return_value={
             "app_source": "main()",
@@ -63,12 +66,14 @@ def test_new_app(
     )
     new_command.tools.cookiecutter = mock.MagicMock(spec_set=cookiecutter)
 
-    # Create the new app, using the default template.
-    new_command.new_app()
+    # Create the new app, using the default template and no project overrides
+    new_command.new_app(project_overrides={})
 
     # App context is constructed
-    new_command.build_app_context.assert_called_once_with(None)
-    new_command.build_gui_context.assert_called_once_with(app_context, None)
+    new_command.build_app_context.assert_called_once_with({})
+    new_command.create_bootstrap.assert_called_once_with(app_context, {})
+    new_command.build_gui_context.assert_called_once_with(bootstrap, {})
+
     # Template is updated
     new_command.update_cookiecutter_cache.assert_called_once_with(
         template="https://github.com/beeware/briefcase-template",
@@ -106,6 +111,8 @@ def test_new_app_missing_template(monkeypatch, new_command, tmp_path):
         "app_name": "myapplication",
     }
     new_command.build_app_context = mock.MagicMock(return_value=app_context)
+    bootstrap = BaseGuiBootstrap(new_command.logger, new_command.input, {})
+    new_command.create_bootstrap = mock.MagicMock(return_value=bootstrap)
     new_command.build_gui_context = mock.MagicMock(
         return_value={
             "app_source": "main()",
@@ -124,11 +131,12 @@ def test_new_app_missing_template(monkeypatch, new_command, tmp_path):
 
     # Create the new app, using the default template. This raises an error.
     with pytest.raises(InvalidTemplateBranch):
-        new_command.new_app()
+        new_command.new_app(project_overrides={})
 
     # App context is constructed
-    new_command.build_app_context.assert_called_once_with(None)
-    new_command.build_gui_context.assert_called_once_with(app_context, None)
+    new_command.build_app_context.assert_called_once_with({})
+    new_command.create_bootstrap.assert_called_once_with(app_context, {})
+    new_command.build_gui_context.assert_called_once_with(bootstrap, {})
 
     # The cookiecutter cache is updated once
     new_command.update_cookiecutter_cache.assert_called_once_with(
@@ -173,6 +181,8 @@ def test_new_app_dev(monkeypatch, new_command, tmp_path, briefcase_version):
         "app_name": "myapplication",
     }
     new_command.build_app_context = mock.MagicMock(return_value=app_context)
+    bootstrap = BaseGuiBootstrap(new_command.logger, new_command.input, {})
+    new_command.create_bootstrap = mock.MagicMock(return_value=bootstrap)
     new_command.build_gui_context = mock.MagicMock(
         return_value={
             "app_source": "main()",
@@ -192,12 +202,14 @@ def test_new_app_dev(monkeypatch, new_command, tmp_path, briefcase_version):
         None,
     ]
 
-    # Create the new app, using the default template.
-    new_command.new_app()
+    # Create the new app, using the default template and no project overrides.
+    new_command.new_app(project_overrides={})
 
     # App context is constructed
-    new_command.build_app_context.assert_called_once_with(None)
-    new_command.build_gui_context.assert_called_once_with(app_context, None)
+    new_command.build_app_context.assert_called_once_with({})
+    new_command.create_bootstrap.assert_called_once_with(app_context, {})
+    new_command.build_gui_context.assert_called_once_with(bootstrap, {})
+
     # Template is updated
     assert new_command.update_cookiecutter_cache.mock_calls == [
         mock.call(
@@ -265,6 +277,8 @@ def test_new_app_with_template(monkeypatch, new_command, tmp_path):
         "app_name": "myapplication",
     }
     new_command.build_app_context = mock.MagicMock(return_value=app_context)
+    bootstrap = BaseGuiBootstrap(new_command.logger, new_command.input, {})
+    new_command.create_bootstrap = mock.MagicMock(return_value=bootstrap)
     new_command.build_gui_context = mock.MagicMock(
         return_value={
             "app_source": "main()",
@@ -276,12 +290,14 @@ def test_new_app_with_template(monkeypatch, new_command, tmp_path):
     )
     new_command.tools.cookiecutter = mock.MagicMock(spec_set=cookiecutter)
 
-    # Create a new app, with a specific template.
-    new_command.new_app(template="https://example.com/other.git")
+    # Create a new app, with a specific template and no project overrides
+    new_command.new_app(template="https://example.com/other.git", project_overrides={})
 
     # App context is constructed
-    new_command.build_app_context.assert_called_once_with(None)
-    new_command.build_gui_context.assert_called_once_with(app_context, None)
+    new_command.build_app_context.assert_called_once_with({})
+    new_command.create_bootstrap.assert_called_once_with(app_context, {})
+    new_command.build_gui_context.assert_called_once_with(bootstrap, {})
+
     # Template is updated
     new_command.update_cookiecutter_cache.assert_called_once_with(
         template="https://example.com/other.git",
@@ -319,6 +335,8 @@ def test_new_app_with_invalid_template(monkeypatch, new_command, tmp_path):
         "app_name": "myapplication",
     }
     new_command.build_app_context = mock.MagicMock(return_value=app_context)
+    bootstrap = BaseGuiBootstrap(new_command.logger, new_command.input, {})
+    new_command.create_bootstrap = mock.MagicMock(return_value=bootstrap)
     new_command.build_gui_context = mock.MagicMock(
         return_value={
             "app_source": "main()",
@@ -335,13 +353,18 @@ def test_new_app_with_invalid_template(monkeypatch, new_command, tmp_path):
         cookiecutter_exceptions.RepositoryNotFound
     )
 
-    # Create a new app, with a specific template.
+    # Create a new app, with a specific template and no project overrides
     with pytest.raises(InvalidTemplateRepository):
-        new_command.new_app(template="https://example.com/other.git")
+        new_command.new_app(
+            template="https://example.com/other.git",
+            project_overrides={},
+        )
 
     # App context is constructed
-    new_command.build_app_context.assert_called_once_with(None)
-    new_command.build_gui_context.assert_called_once_with(app_context, None)
+    new_command.build_app_context.assert_called_once_with({})
+    new_command.create_bootstrap.assert_called_once_with(app_context, {})
+    new_command.build_gui_context.assert_called_once_with(bootstrap, {})
+
     # Template is updated
     new_command.update_cookiecutter_cache.assert_called_once_with(
         template="https://example.com/other.git",
@@ -380,6 +403,8 @@ def test_new_app_with_invalid_template_branch(monkeypatch, new_command, tmp_path
         "app_name": "myapplication",
     }
     new_command.build_app_context = mock.MagicMock(return_value=app_context)
+    bootstrap = BaseGuiBootstrap(new_command.logger, new_command.input, {})
+    new_command.create_bootstrap = mock.MagicMock(return_value=bootstrap)
     new_command.build_gui_context = mock.MagicMock(
         return_value={
             "app_source": "main()",
@@ -398,11 +423,15 @@ def test_new_app_with_invalid_template_branch(monkeypatch, new_command, tmp_path
 
     # Create a new app, with a specific template; this raises an error
     with pytest.raises(InvalidTemplateBranch):
-        new_command.new_app(template="https://example.com/other.git")
+        new_command.new_app(
+            template="https://example.com/other.git",
+            project_overrides={},
+        )
 
     # App context is constructed
-    new_command.build_app_context.assert_called_once_with(None)
-    new_command.build_gui_context.assert_called_once_with(app_context, None)
+    new_command.build_app_context.assert_called_once_with({})
+    new_command.create_bootstrap.assert_called_once_with(app_context, {})
+    new_command.build_gui_context.assert_called_once_with(bootstrap, {})
 
     # Template is updated
     new_command.update_cookiecutter_cache.assert_called_once_with(
@@ -442,6 +471,8 @@ def test_new_app_with_branch(monkeypatch, new_command, tmp_path):
         "app_name": "myapplication",
     }
     new_command.build_app_context = mock.MagicMock(return_value=app_context)
+    bootstrap = BaseGuiBootstrap(new_command.logger, new_command.input, {})
+    new_command.create_bootstrap = mock.MagicMock(return_value=bootstrap)
     new_command.build_gui_context = mock.MagicMock(
         return_value={
             "app_source": "main()",
@@ -454,11 +485,13 @@ def test_new_app_with_branch(monkeypatch, new_command, tmp_path):
     new_command.tools.cookiecutter = mock.MagicMock(spec_set=cookiecutter)
 
     # Create a new app, with a specific template branch.
-    new_command.new_app(template_branch="experimental")
+    new_command.new_app(template_branch="experimental", project_overrides={})
 
     # App context is constructed
-    new_command.build_app_context.assert_called_once_with(None)
-    new_command.build_gui_context.assert_called_once_with(app_context, None)
+    new_command.build_app_context.assert_called_once_with({})
+    new_command.create_bootstrap.assert_called_once_with(app_context, {})
+    new_command.build_gui_context.assert_called_once_with(bootstrap, {})
+
     # Template is updated
     new_command.update_cookiecutter_cache.assert_called_once_with(
         template="https://github.com/beeware/briefcase-template",
@@ -501,6 +534,8 @@ def test_new_app_unused_project_overrides(
         "app_name": "myapplication",
     }
     new_command.build_app_context = mock.MagicMock(return_value=app_context)
+    bootstrap = BaseGuiBootstrap(new_command.logger, new_command.input, {})
+    new_command.create_bootstrap = mock.MagicMock(return_value=bootstrap)
     new_command.build_gui_context = mock.MagicMock(
         return_value={
             "app_source": "main()",
@@ -517,8 +552,11 @@ def test_new_app_unused_project_overrides(
 
     # App context is constructed
     new_command.build_app_context.assert_called_once_with({"unused": "override"})
-    new_command.build_gui_context.assert_called_once_with(
+    new_command.create_bootstrap.assert_called_once_with(
         app_context, {"unused": "override"}
+    )
+    new_command.build_gui_context.assert_called_once_with(
+        bootstrap, {"unused": "override"}
     )
     # Template is updated
     new_command.update_cookiecutter_cache.assert_called_once_with(
@@ -567,6 +605,8 @@ def test_abort_if_directory_exists(monkeypatch, new_command, tmp_path):
         "app_name": "myapplication",
     }
     new_command.build_app_context = mock.MagicMock(return_value=app_context)
+    bootstrap = BaseGuiBootstrap(new_command.logger, new_command.input, {})
+    new_command.create_bootstrap = mock.MagicMock(return_value=bootstrap)
     new_command.build_gui_context = mock.MagicMock(
         return_value={
             "app_source": "main()",
@@ -581,11 +621,13 @@ def test_abort_if_directory_exists(monkeypatch, new_command, tmp_path):
     # Create the new app, using the default template.
     # This will raise an error due to the colliding directory
     with pytest.raises(BriefcaseCommandError):
-        new_command.new_app()
+        new_command.new_app(project_overrides={})
 
     # App context is constructed
-    new_command.build_app_context.assert_called_once_with(None)
-    new_command.build_gui_context.assert_called_once_with(app_context, None)
+    new_command.build_app_context.assert_called_once_with({})
+    new_command.create_bootstrap.assert_called_once_with(app_context, {})
+    new_command.build_gui_context.assert_called_once_with(bootstrap, {})
+
     # Template won't be updated or unrolled
     # Cookiecutter was *not* invoked
     assert new_command.update_cookiecutter_cache.call_count == 0
